CSRF: Play nice with SameSite=Strict :)
SameSite=Strict means we have to be sure and check the existance of both
halves of the double submit token.

Otherwise, the attacker just submits an empty POST half, the browser won't
submit the cookie half (since it won't be coming from the same site), and
the check will pass.