Verify `state` for grant_type=implicit requests.
To verify, we need cookies. That means that, just like for
grant_type=authorization_code, we need to redirect to ?retrieve.

We have to do that redirection in JS, though, since we need to
preserve the #fragment.

Pass `state` in with the redirect so we can do the verification
server-side.